Remove unnecessary casts
Fixes clippy::unnecessary_cast

+++ src/arch/x86_64/kernel/fuse.rs
@@ -654,7 +654,7 @@ fn str_into_u8buf(s: &str, u8buf: &mut [u8]) {
 // TODO: max path length?
 const MAX_PATH_LEN: usize = 256;
 fn str_to_path(s: &str) -> [u8; MAX_PATH_LEN] {
-	let mut buf = [0 as u8; MAX_PATH_LEN];
+	let mut buf = [0; MAX_PATH_LEN];
 	str_into_u8buf(s, &mut buf);
 	buf
 }
